create separate DateTimeRangePicker to allow for import from components package
