+++ ui/components/src/DateTimeRangePicker/index.ts
@@ -11,5 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+export * from './DateTimeRangePicker';
 export * from './TimeRangeSelector';
 export * from './AbsoluteTimePicker';
