--- ui/components/src/TimeRangeSelector/TimeRangeSelector.md
@@ -1,17 +1,0 @@
-# TimeRangeSelector
-
-Date & time selection component to customize what data renders on dashboard. This include relative shortcuts and the ability to pick absolute start and end times.
-
-## Usage
-
-```tsx
-<TimeRangeSelector defaultTimeOption={defaultTimeOption} onChange={setTimeRange} timeOptions={TIME_OPTIONS} />
-```
-
-## Props
-
-| Name               |                                                                              |
-| ------------------ | -----------------------------------------------------------------------| --- |
-| timeOptions        | List of supported shortcuts to show in the select dropdown             |     |
-| defaultTimeOption  | Which predefined relative range to show by default time                |     |
-| TBD                | ...                                                                    |     |
